Removed "unknown" from list of entities (#58), added config options to keep the rainbow formatting and support for disabling translatable text (#56, #57)

+++ src/main/java/com/gmail/artemis/the/gr8/playerstats/config/ConfigUpdateHandler.java
@@ -11,7 +11,7 @@ import com.tchristofferson.configupdater.ConfigUpdater;
 public class ConfigUpdateHandler {
 
     /** Add new key-value pairs to the config without losing comments, using <a href="https://github.com/tchristofferson/Config-Updater">tchristofferson's Config-Updater</a> */
-    public ConfigUpdateHandler(Main plugin, File configFile, int configVersion) {
+    public ConfigUpdateHandler(Main plugin, File configFile, double configVersion) {
         YamlConfiguration configuration = YamlConfiguration.loadConfiguration(configFile);
         updateTopListDefault(configuration);
         configuration.set("config-version", configVersion);
